Fuzz-test the specified functions

This function calls each of the functions it receives with each of the objects specified in what.

Usage

fuzz(
  funs,
  what = test_inputs(),
  package = NULL,
  listify_what = FALSE,
  ignore_patterns = "",
  ignore_warnings = FALSE
)

Arguments

funs

A character vector of function names to test. If a "package" attribute is set, the functions are loaded from that package's namespace; otherwise, they are searched in the global namespace.

what

A list of objects to be passed, one at a time, as the first argument to each function in funs. Ideally, the list should be named, so that each input tested can be pretty-printed with the corresponding name. For unnamed lists, a deparsed representation of the inputs will be used, which may appear unwieldy in some cases. If nothing is provided, a default set of inputs generated by test_inputs() will be used.

package

A character string specifying the name of the package to search for functions. If NULL (default), the function will first check the "package" attribute of funs, and if that is not set, names will be searched in the global namespace.

listify_what

Whether each input in what should also be tested in its listified version (FALSE by default). When set to TRUE, if what is list(x = x), the function will operate as if what were list(x = x, "list(x)" = list(x)), for any input object x.

ignore_patterns

One or more strings containing regular expressions to match the errors to ignore. The string "is missing, with no default" is always ignored.

ignore_warnings

Whether warnings should be ignored (FALSE by default).

Value

An object of class cbtf that stores the results obtained for each of the functions tested.

See also

Examples

## this should produce no errors
res <- fuzz(funs = c("list", "matrix", "mean"),
            what = test_inputs(c("numeric", "raw")))
#> â„¹ Fuzzing 3 functions on 8 inputs
#> â„¹ Functions will be searched in the global namespace as 'package' was not specified
summary(res)
#> Fuzzed 3 functions on 8 inputs:
#>         
#>          FAIL WARN SKIP OK
#>   list      0    0    0  8
#>   matrix    0    0    0  8
#>   mean      0    0    0  8
#> 
#> [ FAIL 0 | WARN 0 | SKIP 0 | OK 24 ]

## this will catch an error (false positive)
fuzz(funs = "matrix",  what = test_inputs("scalar"))
#> â„¹ Fuzzing 1 function on 7 inputs
#> â„¹ Functions will be searched in the global namespace as 'package' was not specified
#> âœ–  ðŸš¨   CAUGHT BY THE FUZZ!   ðŸš¨
#> 
#> â”€â”€ Test input: NULL 
#>  matrix  FAIL  'data' must be of a vector type, was 'NULL'
#> 
#>  [ FAIL 1 | WARN 0 | SKIP 0 | OK 6 ]